VS2008 code fixes

+++ libanalyse/NGram.cpp
@@ -30,11 +30,11 @@ void NGram::Init(int d, int n)
 {
 	ASSERT(d>0 && n>0);
 	m_Rows=d;
-	ASSERT(pow(m_Rows,n-1) < INT_MAX);
-	m_Columns=(int)pow(m_Rows,n-1);
+	ASSERT(pow((double)m_Rows,(double)(n-1)) < INT_MAX);
+	m_Columns=(int)pow((double)m_Rows,(double)(n-1));
 	m_dim=d; m_N=n; 
-	ASSERT(pow(d,n) < INT_MAX);
-	m_hcount.SetSize((int)pow(d,n));
+	ASSERT(pow((double)d,n) < INT_MAX);
+	m_hcount.SetSize((int)pow((long)d,n));
 	//m_totalcount=0;
 	m_Converter=NULL;
 }
